Authentication implementation, App and API

+++ SmartLockerApi/src/main/kotlin/project/smartlocker/http/controller/ModuleController.kt
@@ -1,26 +1,38 @@
 package project.smartlocker.http.controller
 
+import jakarta.servlet.http.HttpServletRequest
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import project.smartlocker.domain.module.ModuleStatus
+import project.smartlocker.domain.user.RoleEnum
 import project.smartlocker.http.models.module.input.CreateModuleRequest
 import project.smartlocker.http.models.module.input.UpdateModuleRequest
 import project.smartlocker.http.models.module.output.AdminModuleDTO
 import project.smartlocker.http.models.module.output.ModuleDTO
+import project.smartlocker.http.models.user.output.UserDTO
 import project.smartlocker.http.utlis.Uris
 import project.smartlocker.services.ModuleService
 
 @RestController
-@RequestMapping(Uris.HOME)
+@RequestMapping(Uris.API)
 class ModuleController(
     private val moduleService: ModuleService
 ) {
     // admin
     @GetMapping(Uris.Module.GET_ALL_MODULES)
-    fun getAllModules(): ResponseEntity<List<AdminModuleDTO>> {
-        val users = moduleService.getAllModules()
-        return ResponseEntity.ok(users)
+    fun getAllModules(
+        request: HttpServletRequest
+    ): ResponseEntity<Any> {
+        val user = request.getAttribute("authenticatedUser") as? UserDTO
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body("Unauthorized: Please log in to access this resource.")
+
+        if (user.role != RoleEnum.ADMIN.toString()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Sorry, you don’t have permission for this page.")
+        }
+        val modules = moduleService.getAllModules()
+        return ResponseEntity.ok(modules)
     }
 
     @GetMapping(Uris.Module.GET_ALL_MODULES_STATUS)
@@ -29,34 +41,91 @@ class ModuleController(
         return ResponseEntity.ok(users)
     }
 
-    @GetMapping(Uris.Module.GET_MODULE_BY_ID)
-    fun getModuleById(@PathVariable id: Int): ResponseEntity<AdminModuleDTO> {
-        val module = moduleService.getModuleById(id)
-        return ResponseEntity.ok(module)
-    }
-
     @PostMapping(Uris.Module.CREATE_MODULE)
-    fun createModule(@RequestBody input: CreateModuleRequest): ResponseEntity<Void> {
+    fun createModule(
+        @RequestBody input: CreateModuleRequest,
+        request: HttpServletRequest
+    ): ResponseEntity<Any> {
+        val user = request.getAttribute("authenticatedUser") as? UserDTO
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body("Unauthorized: Please log in to access this resource.")
+
+        if (user.role != RoleEnum.ADMIN.toString()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Sorry, you don’t have permission for this page.")
+        }
         moduleService.createModule(input)
         return ResponseEntity.status(HttpStatus.CREATED).build()
     }
 
     @PutMapping(Uris.Module.UPDATE_MODULE)
-    fun updateModule(@PathVariable id: Int, @RequestBody input: UpdateModuleRequest): ResponseEntity<Void> {
+    fun updateModule(
+        @PathVariable id: Int,
+        @RequestBody input: UpdateModuleRequest,
+        request: HttpServletRequest
+    ): ResponseEntity<Any> {
+        val user = request.getAttribute("authenticatedUser") as? UserDTO
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body("Unauthorized: Please log in to access this resource.")
+
+        if (user.role != RoleEnum.ADMIN.toString()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Sorry, you don’t have permission for this page.")
+        }
         moduleService.updateModule(id, input)
         return ResponseEntity.ok().build()
     }
 
     @DeleteMapping(Uris.Module.DELETE_MODULE)
-    fun deleteModule(@PathVariable id: Int): ResponseEntity<Void> {
+    fun deleteModule(
+        @PathVariable id: Int,
+        request: HttpServletRequest
+    ): ResponseEntity<Any> {
+        val user = request.getAttribute("authenticatedUser") as? UserDTO
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body("Unauthorized: Please log in to access this resource.")
+
+        if (user.role != RoleEnum.ADMIN.toString()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Sorry, you don’t have permission for this page.")
+        }
         moduleService.deleteModule(id)
         return ResponseEntity.noContent().build()
+
+    }
+    @GetMapping(Uris.Module.GET_MODULE_BY_ID)
+    fun getModuleById(
+        @PathVariable id: Int,
+        request: HttpServletRequest
+    ): ResponseEntity<Any> {
+        val user = request.getAttribute("authenticatedUser") as? UserDTO
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body("Unauthorized: Please log in to access this resource.")
+
+        if (user.role != RoleEnum.ADMIN.toString()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Sorry, you don’t have permission for this page.")
+        }
+        val module = moduleService.getModuleById(id)
+        return ResponseEntity.ok(module)
     }
 
+
+/*
     // global
     @GetMapping(Uris.Module.GET_MODULE_GEO)
     fun getModulesByRadius(@PathVariable latitude: Double, @PathVariable longitude: Double, @RequestParam radius: Double): ResponseEntity<List<ModuleDTO>> {
         val modules = moduleService.getModulesByRadius(latitude, longitude, radius)
         return ResponseEntity.ok(modules)
     }
+
+ */
+
+    // app
+    @GetMapping(Uris.Module.GET_MAP)
+    fun getMap(
+        @RequestParam latitude: Double,
+        @RequestParam longitude: Double,
+        @RequestParam radius: Double
+    ): ResponseEntity<List<ModuleDTO>> {
+        val modules = moduleService.getMap(latitude, longitude, radius)
+        return ResponseEntity.ok(modules)
+    }
+
 }
