+++ SmartLockerApi/src/main/kotlin/project/smartlocker/http/controller/UserController.kt
@@ -1,26 +1,38 @@
 package project.smartlocker.http.controller
 
+import project.smartlocker.http.models.user.output.UserInfoDTO
+import jakarta.servlet.http.HttpServletRequest
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import project.smartlocker.http.models.user.input.CreateUserRequest
-import project.smartlocker.http.models.user.input.UpdateUserRequest
-import project.smartlocker.http.models.user.input.UpdateUserStatusRequest
-import project.smartlocker.http.models.user.output.AdminUserDTO
+import project.smartlocker.domain.user.RoleEnum
+import project.smartlocker.domain.user.User
+import project.smartlocker.http.models.user.input.*
 import project.smartlocker.http.models.user.output.AdminUserStatusDTO
+import project.smartlocker.http.models.user.output.TokenDTO
 import project.smartlocker.http.models.user.output.UserDTO
 import project.smartlocker.http.utlis.Uris
 import project.smartlocker.services.UserService
 
 
 @RestController
-@RequestMapping(Uris.HOME)
+@RequestMapping(Uris.API)
 class UserController(
     private val userService: UserService
 ) {
     // admin
     @GetMapping(Uris.Users.GET_ALL_USERS)
-    fun getAllUsers(): ResponseEntity<List<AdminUserDTO>> {
+    fun getAllUsers(
+        request: HttpServletRequest
+    ): Any {
+        val user = request.getAttribute("authenticatedUser") as? UserDTO
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body("Unauthorized: Please log in to access this resource.")
+
+        if (user.role != RoleEnum.ADMIN.toString()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Sorry, you don’t have permission for this page.")
+        }
+
         val users = userService.getAllUsers()
         return ResponseEntity.ok(users)
     }
@@ -32,10 +44,39 @@ class UserController(
     }
 
     @DeleteMapping(Uris.Users.DELETE)
-    fun deleteUser(@PathVariable id: Int): ResponseEntity<Void> {
+    fun deleteUser(
+        @PathVariable id: Int,
+        request: HttpServletRequest
+    ): Any {
+
+        val user = request.getAttribute("authenticatedUser") as? UserDTO
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body("Unauthorized: Please log in to access this resource.")
+
+        if (user.role != RoleEnum.ADMIN.toString()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Sorry, you don’t have permission for this page.")
+        }
         userService.deleteUser(id)
-        return ResponseEntity.noContent().build()
+        return ResponseEntity.ok("User $id was deleted successfully.")
     }
+
+    @PutMapping(Uris.Users.EDIT_ROLE)
+    fun editUserRole(
+        @PathVariable id: Int,
+        @RequestBody update: UpdateUserRoleRequest,
+        request: HttpServletRequest
+    ): Any {
+        val user = request.getAttribute("authenticatedUser") as? UserDTO
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body("Unauthorized: Please log in to access this resource.")
+
+        if (user.role != RoleEnum.ADMIN.toString()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Sorry, you don’t have permission for this page.")
+        }
+        userService.editRole(id, update)
+        return ResponseEntity.ok("User $id got a new role.")
+    }
+
 
     // global
     @GetMapping(Uris.Users.GET_BY_ID)
@@ -45,33 +86,91 @@ class UserController(
     }
 
     @GetMapping(Uris.Users.GET_BY_USERNAME)
-    fun getUserByUsername(@PathVariable username: String): ResponseEntity<UserDTO>{
+    fun getUserByUsername(
+        @PathVariable username: String,
+        request: HttpServletRequest
+    ): ResponseEntity<UserDTO> {
+        request.getAttribute("authenticatedUser") as? User
+            ?: throw RuntimeException("No user in context")
+
         val user = userService.getUserByUsername(username)
         return ResponseEntity.ok(user)
     }
 
-    @PutMapping(Uris.Users.UPDATE_USER)
-    fun updateUser(@PathVariable id: Int, @RequestBody input: UpdateUserRequest): ResponseEntity<UserDTO> {
-        userService.updateUser(id, input)
+    @PutMapping(Uris.Users.EDIT_STATUS)
+    fun editUserStatus(
+        @PathVariable userId: Int,
+        @RequestBody input: UpdateUserStatusRequest,
+        request: HttpServletRequest
+    ): ResponseEntity<Any> {
+        if (request.getAttribute("authenticatedUser") !is UserDTO) return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+            .body("Unauthorized: Please log in to access this resource.")
+
+        userService.editUserStatus(userId, input)
         return ResponseEntity.ok().build()
     }
 
-    @PutMapping(Uris.Users.VALIDATE_USER)
-    fun validateUser(@PathVariable id: Int, @RequestBody input: UpdateUserStatusRequest): ResponseEntity<UserDTO> {
-        userService.validateUser(id, input)
-        return ResponseEntity.ok().build()
+
+    // app
+    @PostMapping(Uris.Users.LOGIN)
+    fun login(
+        @RequestBody input: LoginRequest
+    ): ResponseEntity<TokenDTO> {
+        val token = userService.login(input)
+        return ResponseEntity.ok(token)
     }
 
-    @PostMapping(Uris.Users.CREATE_USER)
-    fun createUser(@RequestBody input: CreateUserRequest): ResponseEntity<Void> {
+    @PostMapping(Uris.Users.LOGOUT)
+    fun logout(
+        request: HttpServletRequest
+    ): ResponseEntity<Void> {
+        val user = request.getAttribute("authenticatedUser") as? User
+            ?: return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
+
+        userService.logout(user)
+        return ResponseEntity.noContent().build()
+    }
+
+    @PostMapping(Uris.Users.REGISTER)
+    fun register(
+        @RequestBody input: CreateUserRequest
+    ): ResponseEntity<Void> {
         userService.createUser(input)
         return ResponseEntity.status(HttpStatus.CREATED).build()
     }
-}
-/*
-    @PostMapping(Uris.Users.LOGIN)
-    fun login(@RequestBody input: LoginDTO): ResponseEntity<TokenDTO> {
-        val token = userService.login(input.username, input.password)
-        return ResponseEntity.ok(TokenDTO(token))
+
+
+    @GetMapping(Uris.Users.PROFILE)
+    fun getProfile(
+        request: HttpServletRequest
+    ): ResponseEntity<UserDTO> {
+        val user = request.getAttribute("authenticatedUser") as? User
+            ?: throw RuntimeException("No user in context")
+
+        val profile = userService.getUserById(user.id)
+        return ResponseEntity.ok(profile)
     }
-*/
+
+    @PutMapping(Uris.Users.EDIT_USER)
+    fun updateProfile(
+        request: HttpServletRequest,
+        @RequestBody update: UpdateUserRequest
+    ): ResponseEntity<Void> {
+        val user = request.getAttribute("authenticatedUser") as? User
+            ?: throw RuntimeException("No user in context")
+
+        userService.updateUser(user.id, update)
+        return ResponseEntity.noContent().build()
+    }
+
+    @GetMapping(Uris.Users.USER_INFO)
+    fun getUserInfo(
+        request: HttpServletRequest,
+        @PathVariable id: Int
+    ): ResponseEntity<UserInfoDTO> {
+        request.getAttribute("authenticatedUser") as? User
+            ?: throw RuntimeException("No user in context")
+        val user = userService.getUserInfo(id)
+        return ResponseEntity.ok(user)
+    }
+}
