+++ SmartLockerApi/src/main/kotlin/project/smartlocker/services/UserService.kt
@@ -1,13 +1,17 @@
 package project.smartlocker.services
 
+import project.smartlocker.http.models.user.output.UserInfoDTO
 import org.springframework.stereotype.Service
-import project.smartlocker.http.models.user.input.CreateUserRequest
-import project.smartlocker.http.models.user.input.UpdateUserRequest
-import project.smartlocker.http.models.user.input.UpdateUserStatusRequest
+import project.smartlocker.domain.user.RoleEnum
+import project.smartlocker.domain.user.User
+import project.smartlocker.domain.user.UserEnum
+import project.smartlocker.http.models.user.input.*
 import project.smartlocker.http.models.user.output.AdminUserDTO
 import project.smartlocker.http.models.user.output.AdminUserStatusDTO
+import project.smartlocker.http.models.user.output.TokenDTO
 import project.smartlocker.http.models.user.output.UserDTO
 import project.smartlocker.repository.UserRepository
+import java.util.*
 
 @Service
 class UserService(
@@ -25,16 +29,80 @@ class UserService(
     }
 
     fun deleteUser(id: Int){
-        userRepository.deleteUser(id)
+        val user = userRepository.getUserById(id) ?: throw Exception("User not found")
+        if(user != null){
+            userRepository.deleteUser(id)
+        }
+    }
+
+    fun editRole(id: Int, update: UpdateUserRoleRequest){
+        userRepository.getUserById(id) ?: throw Exception("User not found")
+        val role = try {
+            RoleEnum.valueOf(update.role.uppercase())
+        } catch (e: IllegalArgumentException) {
+            throw IllegalArgumentException("Invalid role: ${update.role}")
+        }
+        userRepository.editRole(id, role.name)
     }
 
     // global
-    fun getUserById(id: Int): UserDTO? {
-        return userRepository.getUserById(id)
+    fun getUserByUsername(name:String):UserDTO?{                        //
+        return userRepository.getUserByUsername(name)
     }
 
-    fun getUserByUsername(name:String):UserDTO?{
-        return userRepository.getUserByUsername(name)
+    fun editUserStatus(id: Int, input: UpdateUserStatusRequest){
+        val user = userRepository.getUserById(id) ?: throw Exception("User not found")
+        val status = try {
+            UserEnum.valueOf(input.status.uppercase())
+        } catch (e: IllegalArgumentException) {
+            throw IllegalArgumentException("Invalid role: ${input.status}")
+        }
+
+        when (user.status) {
+            UserEnum.NOT_VERIFIED.toString() -> if (status != UserEnum.VERIFIED) {
+                throw IllegalArgumentException("Can only transition from NOT_VERIFIED to VERIFIED")
+            }
+            UserEnum.VERIFIED.toString() -> if (status != UserEnum.SUSPENDED) {
+                throw IllegalArgumentException("Can only transition from VERIFIED to SUSPENDED")
+            }
+            UserEnum.SUSPENDED.toString() -> if (status != UserEnum.VERIFIED) {
+                throw IllegalArgumentException("Can only transition from SUSPENDED to VERIFIED")
+            }
+        }
+        userRepository.editUserStatus(id, input.status)
+    }
+
+
+
+//--------------------------------------------------------------------------------------------------------
+
+    fun login(input: LoginRequest): TokenDTO {
+        val user = userRepository.getUserByEmail(input.email) ?: throw Exception("Invalid email or password")
+
+        if (input.password != user.password) {
+            throw Exception("Invalid email or password")
+        }
+
+        val newToken = UUID.randomUUID()
+        userRepository.updateUserToken(user.id, newToken)
+
+        return TokenDTO(newToken)
+    }
+
+    fun logout(user: User) {
+         userRepository.deleteToken(user.id)
+    }
+
+    fun createUser(input: CreateUserRequest) {
+        val user = userRepository.getUserByEmail(input.email)
+        if (user != null) {
+            throw Exception("Email already registered")
+        }
+        userRepository.createUser(input.username, input.email, input.password)
+    }
+
+    fun getUserById(id: Int): UserDTO? {
+        return userRepository.getUserById(id)
     }
 
     fun updateUser(id: Int, input: UpdateUserRequest) {
@@ -46,11 +114,7 @@ class UserService(
         userRepository.updateUser(id, newEmail, newPassword)
     }
 
-    fun validateUser(id: Int, input: UpdateUserStatusRequest){
-        userRepository.validateUser(id, input.status )
-    }
-
-    fun createUser(newUser: CreateUserRequest) {
-       userRepository.createUser(newUser.username, newUser.email, newUser.password)
+    fun getUserInfo(id: Int): UserInfoDTO? {
+        return userRepository.getUserInfo(id)
     }
 }
